+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Registered output filters wouldn't run [#899](https://github.com/smarty-php/smarty/issues/899)
+
+### Removed
+- Removed `$smarty->registered_filters` array
+
 ## [5.0.0-rc1] - 2023-08-08
 
 ### Added